# docs/swagger.yaml
basePath: /
definitions:
...
  dto.PersonalLetterResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      letter_date:
        type: string
      scan_image_path:
        type: string
      title:
        type: string
    type: object
...
  handler.JSONResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: API for Hemra Şirow literary and biography website
  title: Literary Backend API
  version: "1.0"
paths:
...
  /api/personal-letters:
    get:
      description: List all personal letters
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PersonalLetterResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: List personal letters
    post:
      consumes:
      - application/json
      - multipart/form-data
      description: Create a new personal letter. Supports JSON or multipart/form-data
        with optional scan upload.
      parameters:
      - description: Title
        in: formData
        name: title
        required: true
        type: string
      - description: Content
        in: formData
        name: content
        required: true
        type: string
      - description: Letter date in RFC3339
        in: formData
        name: letter_date
        type: string
      - description: Scan image or PDF
        in: formData
        name: scan
        type: file
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PersonalLetterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Create personal letter
  /api/personal-letters/{id}:
    delete:
      description: Delete an existing personal letter by id
      parameters:
      - description: Personal letter ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Delete personal letter
    get:
      description: Get a single personal letter by id
      parameters:
      - description: Personal letter ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PersonalLetterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Get personal letter
    put:
      consumes:
      - application/json
      - multipart/form-data
      description: Update an existing personal letter. Supports JSON or multipart/form-data
        with optional scan upload.
      parameters:
      - description: Personal letter ID
        in: path
        name: id
        required: true
        type: string
      - description: Title
        in: formData
        name: title
        type: string
      - description: Content
        in: formData
        name: content
        type: string
      - description: Letter date in RFC3339
        in: formData
        name: letter_date
        type: string
      - description: Scan image or PDF
        in: formData
        name: scan
        type: file
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PersonalLetterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Update personal letter
...
schemes:
- http
- https
swagger: "2.0"